Fix credits in footer of each article
Aparently, if $wgMaxCredits, the ->getCredits() method of the 'credits' action
is called by SkinTemplate to fill the footer.
1/ It did not call that method with the parameter (page id) the method expected
2/ We already had a hook in place, so we didn't really want that

However, since the other method is built-in and it makes no sense to have a
hook in there that we don't even need, this patch gets rid of the hook, changes
the method to accept the parameters SkinTemplate will provide, and take care of
it from there.

Change-Id: I29ba341ab8dc4717aa0a9558b2be2e9f2c0aeea7